fix companion app syncing

+++ src/main/kotlin/com/plugin/features/auth/github/DTO.kt
@@ -9,9 +9,9 @@ import com.fasterxml.jackson.annotation.JsonProperty
 @JsonInclude(JsonInclude.Include.NON_NULL)
 data class GitHubOAuthTokenResponse(
     @JsonProperty("access_token") val accessToken: String,
-    @JsonProperty("expires_in") val expiresIn: Int,
+    @JsonProperty("expires_in") val expiresIn: Long,
     @JsonProperty("refresh_token") val refreshToken: String,
-    @JsonProperty("refresh_token_expires_in") val refreshTokenExpiresIn: Int,
+    @JsonProperty("refresh_token_expires_in") val refreshTokenExpiresIn: Long,
     @JsonProperty("scope") val scope: String = "",
     @JsonProperty("token_type") val tokenType: String = "bearer",
 )
